Move tests to same file to fix CI
